MQTTInspectorView: lift JSONHighlighter and SubscribeStore into peer files
MQTTInspectorView.swift was 411 lines holding the inspector view, an
@Observable store, a model type, a generic JSON highlighting utility, and
two helper subviews. Splits:

- JSONHighlighter (a generic regex-based JSON colorer with no MQTT-specific
  state) → Shared/Components/JSONHighlighter.swift, where it's reusable
  by anything else surfacing JSON to the user (BeautifulPayloadView is the
  obvious next consumer). Pure Foundation/SwiftUI, so no widget membership
  exception needed.
- SubscribeStore + InspectorMessage (the @Observable buffer of inbound
  messages plus the per-message model that decodes the topic into a log
  level for coloring) → Features/Settings/Developer/SubscribeStore.swift.
  Tightly coupled, so they stay together; both reference main-app types
  (ConnectionSessionController, JSONValue, Z2MTopics, LogLevel) and need
  the widget membership exception.

The view file itself drops from 411 → 276 lines and now holds only the
view layer (MQTTInspectorView, SubscribeView, MessageRow, PublishView).

Refs #33
